+++ src/main/java/edu/harvard/iq/dataverse/api/Access.java
@@ -206,13 +206,7 @@ public class Access extends AbstractApiBean {
         DownloadInfo dInfo = new DownloadInfo(df);
         BundleDownloadInstance downloadInstance = new BundleDownloadInstance(dInfo);
 
-        FileMetadata fileMetadata = null;
-
-        if (fileMetadataId == null) {
-            fileMetadata = df.getFileMetadata();
-        } else {
-            fileMetadata = dataFileService.findFileMetadata(fileMetadataId);
-        }
+        FileMetadata fileMetadata = getFileMetadataToExport(df, fileMetadataId, req);
         
         downloadInstance.setFileCitationEndNote(new DataCitation(fileMetadata).toEndNoteString());
         downloadInstance.setFileCitationRIS(new DataCitation(fileMetadata).toRISString());
@@ -220,14 +214,13 @@ public class Access extends AbstractApiBean {
 
         ByteArrayOutputStream outStream = null;
         outStream = new ByteArrayOutputStream();
-        Long dfId = df.getId();
         try {
             ddiExportService.exportDataFile(
-                    dfId,
+                    df,
                     outStream,
                     null,
                     null,
-                    fileMetadataId);
+                    fileMetadata);
 
             downloadInstance.setFileDDIXML(outStream.toString());
 
@@ -675,33 +668,14 @@ public class Access extends AbstractApiBean {
         DataverseRequest req = createDataverseRequest(getRequestUser(crc));
         dataFile = findDataFileUserCanSeeOrDieWrapper(fileId, req);
         
+        // This will throw a ForbiddenException if access isn't authorized:
+        checkAuthorization(req.getUser(), dataFile);
+
         if (!dataFile.isTabularData()) { 
            throw new BadRequestException("tabular data required");
         }
-        if (FileUtil.isRetentionExpired(dataFile)) {
-            throw new BadRequestException("unable to download file with expired retention");
-        }
-        if (dataFile.isRestricted() || FileUtil.isActivelyEmbargoed(dataFile)) {
-            boolean hasPermissionToDownloadFile = false;
 
-            if (req != null && req.getUser() instanceof GuestUser) {
-                // We must be in the UI. Try to get a non-GuestUser from the session.
-                req = dvRequestService.getDataverseRequest();
-            }
-            hasPermissionToDownloadFile = permissionService.requestOn(req, dataFile).has(Permission.DownloadFile);
-            if (!hasPermissionToDownloadFile) {
-                throw new BadRequestException("no permission to download file");
-            }
-        }
-
-        response.setHeader("Content-disposition", "attachment; filename=\"dataverse_files.zip\"");
-
-        FileMetadata fm = null;
-        if (fileMetadataId == null) {
-            fm = dataFile.getFileMetadata();
-        } else {
-            fm = dataFileService.findFileMetadata(fileMetadataId);
-        }
+        FileMetadata fm = getFileMetadataToExport(dataFile, fileMetadataId, req);
 
         String fileName = fm.getLabel().replaceAll("\\.tab$", "-ddi.xml");
 
@@ -710,14 +684,13 @@ public class Access extends AbstractApiBean {
         
         ByteArrayOutputStream outStream = null;
         outStream = new ByteArrayOutputStream();
-        Long dataFileId = dataFile.getId();
         try {
             ddiExportService.exportDataFile(
-                    dataFileId,
+                    dataFile,
                     outStream,
                     exclude,
                     include,
-                    fileMetadataId);
+                    fm);
 
             retValue = outStream.toString();
 
@@ -2289,6 +2262,34 @@ public class Access extends AbstractApiBean {
             throw new ForbiddenException();
         }        
     }
+
+    private FileMetadata getFileMetadataToExport(DataFile dataFile, Long fileMetadataId, DataverseRequest req) {
+        FileMetadata fm = null;
+        if (fileMetadataId == null) {
+            if (permissionService.requestOn(req, dataFile.getOwner()).has(Permission.ViewUnpublishedDataset)) {
+                fm = dataFile.getFileMetadata();
+            } else {
+                try {
+                    fm = dataFile.getLatestPublishedFileMetadata();
+                } catch (UnsupportedOperationException e) {
+                    throw new ForbiddenException("No released version found");
+                }
+            }
+        } else {
+            fm = dataFile.getFileMetadatas().stream()
+                    .filter(fileMetadata -> fileMetadataId.equals(fileMetadata.getId()))
+                    .findFirst()
+                    .orElse(null);
+            if (fm == null) {
+                throw new BadRequestException("Invalid fileMetadataId for the given fileId");
+            }
+            if (!fm.getDatasetVersion().isReleased() && !permissionService.requestOn(req, dataFile.getOwner()).has(Permission.ViewUnpublishedDataset)) {
+                throw new ForbiddenException("No permission to view unpublished metadata");
+            }
+        }
+        return fm;
+    }
+
     private User getRequestor(User user) {
         // CompoundAuthMechanism should find the user by API Key/Token, Workflow, etc. And for SPA the Bearer Token
         // For JSF check if CompoundAuthMechanism couldn't find the user then try to get it from the session